Model added to run solve problem 2.
This model will be started with a script and run through the dynamics/kinatics
of the model. It will call varies matlab functions. so far it will call the path functions for x,y over time and pass it in to an IK to get the joint positions. Jacobians have to be added to get the joint velocities.