Several files were added. Two additional models wre tested.
main_1. and prolem2_1.slx do not work, throw algerbra error.

main_3.m, and problem2_3.slx uses only joint space. The joint kinmatics are found pior to the start of the model.
This is done by planning in joint space not in work space. Thus, the trajecories are in joint angles. This by passes the need to use jacobians.
After some tuning of te PI values the model displayed the correct values.